requirements.txt
flask
requests
python-dotenv
sentence-transformers
faiss-cpu
transformers==4.35.0
torch
pandas
flask-cors
openai
